requirements.txt
flask
pytesseract
openai
python-dotenv